+++ ast/src/main/kotlin/com/strumenta/kolasu/ast/Source.kt
@@ -1,0 +1,7 @@
+package com.strumenta.kolasu.ast
+
+abstract class Source : Comparable<Source> {
+    protected abstract fun stringDescription(): String
+
+    override fun compareTo(other: Source): Int = this.stringDescription().compareTo(other.stringDescription())
+}
